// src/infrastructure/interfaces/statistic/ip-data.interface.ts
export interface IIpData {
  ip: string;
  latitude: number;
  longitude: number;
  countryName: string; // Kazakhstan
  cityName: string; // Almaty
  timezoneName: string; // Country/Zone
  fullAgentName: string;
  locales: string[];
  defaultLocale: string; // en_GB
  languageCode: string; // en
  isMobile: boolean;
}